# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ReWear API Documentation",
        "description": "ReWear - Sustainable Fashion Marketplace API\n *\n * **Mission:** Making fashion sustainable and accessible\n *\n * **Tagline:** Buy, Sell, Donate\n *\n * ## Features\n * - 🛍️ **Marketplace:** Buy and sell pre-loved clothing\n * - 💚 **Donations:** Donate items to verified charities\n * - 🚚 **Delivery:** Local driver network\n * - 🌍 **Impact:** Track environmental and social impact\n *\n * ## Authentication\n * All protected endpoints require a JWT Bearer token in the Authorization header:\n * ```\n * Authorization: Bearer YOUR_JWT_TOKEN\n * ```\n *\n * ## Rate Limiting\n * - **Registration OTP:** 5 requests per email per hour, 5 per IP per hour\n * - **Login:** Standard Laravel rate limiting\n * - **API Calls:** 60 requests per minute\n *\n * ## User Roles\n * - **user:** Regular buyer/seller (default)\n * - **charity:** Organization receiving donations (admin-created)\n * - **driver:** Verified delivery personnel (admin-approved)\n * - **admin:** Platform administrator\n *\n * ## Support\n * - **Email:** support@rewear.com\n * - **Phone:** +961 70 123456\n * ",
        "contact": {
            "name": "ReWear Support Team",
            "email": "support@rewear.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
...
        },
...
    ],
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with email and password",
                "description": "Login with email and password",
                "operationId": "7b9026fc5982d72a0a0978e3c868fc33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePassword123!"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "iPhone 13"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials or account locked"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/login-code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request verification code for login (optional)",
                "description": "Request login verification code (optional OTP for login)",
                "operationId": "44e019ea8c232dde06167e3c169ede52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification code sent"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user profile",
                "description": "Get current authenticated user",
                "operationId": "260b31ba54def17299a60027263986d4",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/profile": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile",
                "description": "Update user profile",
                "operationId": "bf5db0ea9448d0bb2797e1c1560daf01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+96170123456"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Fashion enthusiast"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Beirut"
                                    },
                                    "location_lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.8886
                                    },
                                    "location_lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35.4955
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Change user password",
                "description": "Change password",
                "operationId": "1f548e87771d4850558c7c6eb001b5b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "OldPassword123!"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword456!"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword456!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully"
                    },
                    "422": {
                        "description": "Validation error or incorrect old password"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}